fix name path explorer and rename component explorer to index
