+++ client/src/main.tsx
@@ -10,7 +10,7 @@ import App from "./App";
 import Artists from "./pages/Artists";
 import ArtistDetails from "./pages/Artists/ArtistsDetails";
 import CoupDeCoeur from "./pages/CoupDeCoeur";
-import Explorer from "./pages/Explorer/explorer";
+import Explorer from "./pages/Explorer";
 import Home from "./pages/home";
 import Podcasts from "./pages/podcasts";
 import { authLoader } from "./services/Auth/loader";
